backend/agent_service/requirements.txt
# requirements.txt
fastapi>=0.100.0
uvicorn>=0.22.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
openai>=1.0.0
langchain>=0.0.267
langchain-core
langchain-openai
langchain-community
python-dotenv>=1.0.0
asyncio
aiohttp
pylint
pytest
httpx
python-multipart